Avoid running the security pipeline on every PR (#12)
The security pipeline defined in
`azure_pipelines/code-security-analysis.yml` is scheduled to run weekly
as required by Microsoft policy. However, it also seems to run on all
PRs, which is unnecessary. This PR changes the pipeline's configuration
to only do the scheduled runs and not run as part of CI.