Improve the packing number heuristic used for estimating diversity (#22)
To estimate the diversity of a set of routes, we currently use a
heuristic packing number algorithm which finds the largest set of routes
that are sufficiently different from each other (e.g. use
non-overlapping reactions).

In this PR, I propose using an improved algorithm for 10 trials instead
of the current one for 100 trials. Empirically this is ~2x faster while
finding slightly larger packing sets (~1 extra route in the best packing
sets if the input is sufficiently challenging e.g. contains a packing
set with 12 or more routes).

---------

Co-authored-by: Austin Tripp <ajt212@cam.ac.uk>