Add a top-level CLI for running end-to-end search (#26)
This PR adds a first version of a top-level CLI for running search. It
supports all single-step models integrated so far, as well as both MCTS
and Retro* as choices for the search algorithm. Both of these algorithms
have reasonable defaults set; for now these don't depend on the
single-step model, but after tuning I will add model-specific defaults.