Fix error in graph standardization (#24)
The initial release included a file
`syntheseus/search/graph/standardization.py` which contained code to
turn all types of graphs into minimal AND/OR graphs (i.e. AND/OR graphs
with one node per reaction and one node per molecule).
It turns out that for MolSetGraph objects the code to fetch the reaction
objects was incorrect and would crash. I didn't catch this before
because I didn't actually write tests for this file.

This PR fixes this bug, and adds minimal "smoke" tests to make sure that
this method runs without crashing. However, the actual output is not
tested. To reflect this, the method still raises a warning when it is
called.

---------

Co-authored-by: Krzysztof Maziarz <krzysztof.maziarz@microsoft.com>