Add a time limit multiplier to all algorithm base tests (#25)
Many algorithm tests involve running the algorithms for a fixed number
of iterations or maximum time limit, then checking that a certain
outcome happened. However, algorithms with a slow iteration speed may
not achieve the desired outcome within the time limit if their iteration
time is slow. I found an example of this when writing an algorithm that
does a lot of computation with numpy arrays.

If we want slower algorithms to be testable by extending the
`BaseAlgorithmTest`, I saw two options:

1. Extend the tolerances of all tests in the base class. This would have
the side effect of making the tests for all algorithms take longer (e.g.
5s -> 30s)
2. Add a "time multiplier" which defaults to a small value, but can be
overridden by child classes that want more time.

I decided that option 2 makes more sense, so this PR implements it.
Specifically, I added an attribute `time_limit_multiplier` to
`BaseAlgorithmTest`, and replaced all instances of `time_limit_s=X` with
`time_limit_s=X*self.time_limit_multiplier` (except in the test for time
limits themselves because there an absolute tolerance is tested). This
can be overridden in child classes.

---------

Co-authored-by: Krzysztof Maziarz <krzysztof.maziarz@microsoft.com>