Speed up the diversity estimation heuristic (#28)
This PR applies three small tweaks to the diverse subset extraction
heuristic, speeding it up without changing semantics:
- The recursive calls of `_recursive_construct_packing_set` all make use
of the same (unchanging) list of routes, just consider different slices
of that list. Instead of slicing the input list repeatedly in each
recursive call (which does a lot of copying), it is more efficient to
just pass in indices pointing to the slice under consideration (this
changes the signature of the private `_recursive_construct_packing_set`,
but the signature of the public `estimate_packing_number` remains
unchanged).
- The `_jaccard_distance` function was instantiating both a set
intersection and a set union of its arguments, whereas intersection size
alone is enough to compute the distance.
- With `n <= 1` as a base case, `_recursive_construct_packing_set` was
spending a lot of time on finding diverse subsets for inputs with tiny
`n` e.g. `n = 2`, as there is overhead in e.g. creating tiny graphs. It
is more efficient to make `n = 2` one of the base cases, as it's trivial
to solve it optimally.

Empirically, these changes speed up the algorithm by 3%, 3% and 22%,
respectively (together compounding to a speedup of around 26%).