Guard against rare issues in LocalRetro (#31)
In very rare cases LocalRetro's `get_k_predictions` can fail with
`RuntimeError` originating from `rdkit`. I have not seen this with the
USPTO-trained checkpoint, but it did happen once when evaluating a
Pistachio-trained checkpoint (which has a much larger template
diversity). This PR guards against this by simply falling back to
returning no results if an error is raised.