Refactor single-step evaluation script and move it to cli/ (#43)
This PR includes a bit of cleanup and refactoring of the single-step
evaluation script:
- Moved some parts (especially those imported by other files) out to
separate utils files, in order to keep the length of the script a bit
more manageable (the version currently on `main` has ~550 lines).
- Dropped the "combined" metric, which computed accuracy with success
defined as "either the back-translation is successful _or_ the output
matches the ground-truth". This was reported _on top of the individual
back-translation and dataset-based accuracies_; after further thought I
think such a combined metric is a bit complicated and overall adds
clutter.
- Moved the refactored script from `syntheseus/reaction_prediction/cli`
to `syntheseus/cli`, to make sure all CLI endpoints are equally
accessible no matter which part of the underlying functionality they
focus on. The associated `eval.md` documention I moved to a new `docs/`
directory, which will later form the basis for building static
documentation.

Finally, I also fixed a typo in `README.md`.