Add a general CLI endpoint (#44)
Currently, the CLIs defined in `cli/` are not installed as part of the
`pip`-based setup, so the only way to run them is to launch the script
directly from source. This PR changes that by adding a top-level CLI
endpoint, which then allows to select from `search` and
`eval-single-step`.