Add initial documentation (#45)
This PR adds an initial static documentation built using Material for
MkDocs. The docs are built automatically using GitHub Actions and
deployed to microsoft.github.io/syntheseus/.

This first version is a combination of content we had in `README.md`
(which is changed here to be very minimalistic, linking to the docs for
further information) and the pre-existing
`docs/cli/eval_single_step.md`. In future PRs we can extend the docs
with documentation for the search CLI, as well as an API reference and
tutorials.