Automatically download a default checkpoint if it is not provided (#48)
This PR extends the single-step model classes to automatically download
the default (trained on USPTO-50K) model checkpoint if no model
directory is provided. This makes it even easier to get started with
`syntheseus` as one does not have to manually download and unzip the
files. Once downloaded, the checkpoints are cached.

On top of this, I also extend the integration CI to run a new
single-step model test that checks that the supported models do run and
return reasonable predictions for a simple input. While locally the test
works for all 6 `pip`-installable models, on GitHub the GPU is not
available, and setting `device="cpu"` uncovers that some of the models
appear to only work on GPU. Thus, for now the test only covers
Chemformer, LocalRetro and MEGAN; the other models will be added in the
future.

When running those tests, I found that the links we provided for
LocalRetro and RetroKNN were not quite correct (the initial upload
missed some of the files, which I later corrected, but the links were
still pointing to the first upload rather than the corrected one).
Morever, I found that RetroKNN was zipped using a method that `python`
could not deal with, so I re-generated the file and re-uploaded it
(again fixing the link).