Get all models to work on CPU (#57)
PR #48 added a new CI pipeline that covers single-step model inference,
but it only tested half of the single-step models as the other ones did
not work on CPU. This PR fixes this, making all models work on CPU,
which mostly required making sure the right device is correctly
propagated into all of the model internals. I then extended the test
which now checks all `pip`-installable models we support.

Additionally, I also merged the two CI pipelines into one, as this
allows for generating a combined coverage report that takes into account
all tests. This does not affect testing speed, as the different test
jobs still run in parallel. Finally, I also noticed that the
`.coveragerc` configuration was not quite right, as the files with 0%
coverage were not being discovered; after a minor change to the
configuration this is now fixed.