Remove PredictionList (#61)
This PR replaces `PredictionList` objects (and their subclasses) with
`Sequence[Prediction]`. It is meant as the first step to merge the two
kinds of reaction model interfaces. The `PredictionList` class was then
deleted.

I chose `Sequence[Prediction]` instead of `list[Prediction]` to allow
`PredictionList`-like functionality in the future if we ever desire it.
This would be implemented by subclassing `Sequence`.

I also changed the search `BackwardReactionModel` class to return
`list[Sequence]` instead of `list[list]` to prepare for the class merger
in a subsequent PR.

---------

Co-authored-by: Krzysztof Maziarz <krzysztof.maziarz@microsoft.com>