Add new compromise reaction class (#63)
This PR adds a new compromise Reaction class in a new file:
`syntheseus/interface/reaction.py`. The class has the features we
discussed in #58 [^1]. The class also includes a basic test.

The new classes is not used anywhere at the moment[^2]. I will refactor
the rest of the code to use it in another PR. I thought it would be test
to propose the class in isolation first to avoid doing all the
refactoring, then re-doing it if we decide to make additional changes.

[^1]: Except for the `unique_reactants` alias. After thinking about it a
bit more, if users know to use `unique_reactants` then they could just
as easily write `set(reactants)`.

[^2]: Except that I moved the `REACTION_SEPARATOR` constant to
`reaction.py` and changed resulting imports. It seemed odd to have it in
`molecule.py`