Integrate the Graph2Edits model (#65)
This PR integrates Graph2Edits - a recently proposed model class which
shows promising results on USPTO-50K. It includes a full model wrapper
with a default checkpoint, as well as edits to docs and tutorials to
mention the new model type.

Testing the integration uncovered a potential problem with some of the
model classes: as the underlying repositories are often not set up for
installability and use relative imports, creating LocalRetro loads a
Python module called `models`; when one then tries to create Graph2Edits
which also contains a module with that name, Python looks up the module
in `sys.modules` rather than re-importing it from the new location. This
leads to errors if the two models are used together, which was caught by
our models test. To fix this, I refactored LocalRetro to place all
modules under a top-level directory named `local_retro`, and released
the updated code as a new version of `syntheseus-local-retro` on PyPI.

Finally, as import-related errors could potentially depend on the
_order_ in which the two conflicting models are used, I also changed the
models test to test each model type _twice_, covering _each pair of
models in both orders_. Note that this does not affect running time, as
most of the time is spent downloading model weights, and so re-testing a
model again requires negligible extra time.