Improve retro* implementation (#62)
Improved retro* implementation in several ways:

1. **Use better priority queue implementation**: the existing
implementation did not track the set of nodes in the queue, and
therefore could add a node to the queue multiple times. This generally
caused the queue to be much larger than necessary, creating considerable
slowdown in long searches. I fixed this by using a simple PQ
implementation which allows for re-insertion.
2. **Track minimum cost route found so far**: this is essentially the
"reaction number", but disregarding the search heuristic function.
Tracking this could be useful for learning from self-play, and can be
used to check whether the algorithm has found the minimum-cost route
when the heuristic is admissible.
3. **Improve logging**: the queue size, graph size, and other details
are logged in each iteration. I found this useful when running retro*
for a large number of iterations.
4. **Fix edge case bug** where `reaction_number` depends on the
`reaction_number_estimate` for all leaf nodes, including those which
cannot be expanded (e.g. because they are too deep). This could cause
the algorithm to pursue synthesis routes whose estimated cost can never
be realized because one (or more) leaf nodes can never be expanded. I
fixed this by only filling the `reaction_number_estimate` of nodes which
can be expanded, and only including the `reaction_number_estimate` in
the `reaction_number` calculation if it is present.

---------

Co-authored-by: Krzysztof Maziarz <krzysztof.maziarz@microsoft.com>