Add numpy 2 compatibility check to ruff (#71)
Numpy 2.0 will be released soon and it is advised to check all packages
which depend on numpy to ensure that they won't break when this happens
(as there are a few API changes). Thankfully there is a ruff check for
this. I added it to our `pyproject.toml` and bumped the ruff version in
our pre-commit check.