Fix bugs recently introduced into the single-step eval script (#72)
A recent refactoring PR #67 introduced some bugs into
`eval_single_step.py`, which evaded all of our CI and tests, and made
the script no longer run. This PR is a quick (and somewhat dirty) fix to
get the script to run again. Future PRs will refactor the source of the
confusion that gave rise to these bugs, and also improve test coverage
to avoid such issues in the future.