Simplify class hierarchy for reaction objects (#74)
In #63 the new reaction class used the attribute "product" (no 's') to
refer to both the single product in `SingleProductReaction` and the set
of products in `MultiProductReaction`s. This is a bit confusing, so this
PR refactors the class structure so that `MultiProductReaction` is
simply called `Reaction`, and `SingleProductReaction` is a subclass
which has a `product` property. The resulting class structure contains no
generics and is much simpler.