Improve test coverage and error handling in single-step evaluation (#75)
Recently we had some breaking changes to `eval_single_step.py` merged to
`main` due to insufficient test coverage. This PR extends the coverage
by adding tests for `compute_metrics` (using both a forward and a
backward reaction model). It also improves error handling in the
evaluation script (errors otherwise produced by `OmegaConf` tend to be a
bit cryptic) and allows `get_model` to make use of default model weights
(otherwise this feature was not accessible when running through one of
the the CLI entrypoints).