Enable docs versioning (#77)
So far our documentation was not versioned, meaning that the website
always served the version that was most recently built. However, we
generally recommend installing the latest `pip`-released version, which
led to a pitfall in which we release new features or backward-compatible
changes to `main` but delay updating the docs because that would make
them out-of-sync with the state of the library as available via `pip`.

This PR resolves this issue by enabling docs versioning using `mike`.
Now the docs are still built regularly (automatically on each merge to
`main`, but this can also be triggered manually) but these builds
overwrite a special version called `dev`, which is supposed to
correspond to the state of `main`. On top of this, there are versions of
the docs corresponding to `pip`-released versions of the package, which
are released manually and shouldn't be overwritten. This is done via
```
mike deploy --push --update-aliases x.y.z stable
```
The above releases `x.y.z` and also reset the version alias `stable`;
this should be set to the latest `pip`-released version.

Both the CI pipeline as well as local calls to `mike` generate new
commits on the `gh-pages` branch. Currently, two versions are released:
- `dev`, which was released by triggering the CI pipeline
- `0.3.0`, which I released manually from an older version of the
code matching the `v0.3.0` release