Add tutorial for integrating custom models and improve README (#79)
This PR adds a new tutorial that shows how to integrate a custom
reaction model and run search with it (for demonstration purposes it
uses a toy model which is a simplified version of a model we use in
tests). Additionally, it also slightly improves the README by adding
links at the top to docs, CLI reference, tutorials and our paper. The
first three technically all link to somewhere in the docs, but having
them spelled out at the top may make it more clear to users that these
things (especially tutorials) do exist. Finally, I also apply a small
cleanup to `.gitignore` to remove an entry which is no longer relevant.