Update PaRoutes search tutorial (#82)
This PR replaces the old search tutorial (which lives under
`tutorials/search`) with an updated tutorial which lives under
`docs/tutorials`. The content is mostly the same as the old tutorial,
although I merged the two parts into 1. Also, instead of including the
code to run the PaRoutes benchmark as a `.py` file in the same directory
as this tutorial, I made it into a separate python package (available on
[GitHub](https://github.com/AustinT/syntheseus-paroutes) and
[PyPI](https://pypi.org/project/syntheseus-paroutes/)).

---------

Co-authored-by: Krzysztof Maziarz <krzysztof.maziarz@microsoft.com>