Fix small bug in AND/OR graph validity check (#84)
One of the checks ensures that the number of children of an AND node
matches the number of reactants in its reaction. When the type of
`AndNode.reaction.reactants` was changed from `frozenset` to `Bag`, this
check should have been updated to check `unique_reactants` instead of
`reactants`. This PR fixes this and adds a more informative error
message.