Fix incorrectly uploaded RetroKNN weights (#91)
As noticed by @panpiort8, the default USPTO50K-trained RetroKNN weights
did not reproduce the USPTO50K accuracy reported in our paper, only
getting to ~50% instead of ~55%. It turns out that the weights got mixed
up when uploading from our internal storage to figshare. In short, the
RetroKNN weights consist of LocalRetro weights and an activation store
for retrieval, but the store present in the checkpoint was generated
using a _different_ checkpoint of LocalRetro than the one it was
uploaded with. After fixing this, the accuracy is back to its expected
~55% range.