Fix issues with GLN (#92)
The GLN model is the only model class that uses a bespoke environment
based on Python 3.7 instead of the shared environment that uses Python
3.9. Due to this, it is not tested in CI, and is generally prone to bugs
being introduced over time.

After trying to run the GLN model using the weights uploaded to
figshare, I found that the upload was incomplete, and some files were
missing from the checkpoint, preventing the model from being loaded. On
top of this, there were a few small issues in the wrapper that
accumulated over time, which also needed to be addressed to get the
model to run again. This PR changes the weights link to a complete
upload and also fixes the small issues in the model wrapper.