Fix constructor arguments in `ParallelReactionModel` (#96)
One of the more rarely used capabilities of our codebase is running
multi-GPU inference during single-step evaluation. Testing this part of
code requires having a GPU, and so `ParallelReactionModel` was not
covered by unit tests, and it broke at some point during refactoring of
the model classes. This PR adds simple tests for it and fixes the issue.