Fix coverage report upload in the CI (#101)
Earlier this week, GitHub rolled out an update to the `upload-artifact`
action, which makes it ignore hidden files by default (this is to reduce
the risk of accidentally including unwanted files e.g. credentials). As
the `.coverage.*` files are hidden (i.e. their name starts with `.`),
they started being skipped in the upload, and our CI started failing.
This PR fixes things by setting the `include-hidden-files` flag to
restore the previous behaviour.