Expose `stop_on_first_solution` as a CLI flag (#100)
In #13, a `stop_on_first_solution` flag was added to the base search
algorithm class, but so far it was not available when using the search
CLI. As this option can be quite useful (for e.g. running a quick check
to assess the difficulty of a given set of targets), this PR adds it as
a CLI option as well.