Extend single-step evaluation with stereo-agnostic results (#102)
This PR adds an `include_results_modulo_stereo` flag to the single-step
evaluation script to additionally compute metrics (top-k accuracy, MRR)
under a relaxed notion of success which ignores stereochemistry
information. This can be useful to assess what fraction of errors made
by the model are due to a mismatch in chirality. Additionally, this PR
also slightly tidies up the printing of eval results.