Update GLN environment to Python 3.9 (#103)
Unlike all other supported single-step models, GLN uses a separate
environment defined with docker, as it depends on building rdkit from
source. GLN is also the only model using Python 3.7 - a version which
reached its end-of-life last year. This PR keeps the docker-based setup,
but updates the GLN code and environment to Python 3.9, so that in a
future version of syntheseus we can raise the required Python version to
3.9 and get rid of backward compatibility hacks.