Pin further Chemformer dependencies to avoid `torch` reinstallation (#108)
Recently, the CI stopped working, as `pip`-installing `syntheseus[all]`
started triggering a reinstallation of `torch`, despite the fact that we
explicitly set up the right version of `torch` in the base environment.
This was caused by `torchmetrics` having been recently updated to 1.6.0;
despite the fact that `syntheseus-chemformer` pins `pytorch-lightning`,
the latter seemingly doesn't pin `torchmetrics`, as we started getting
the new version after it was released, which in turn is not compatible
with the version of `torch` we use, prompting a reinstallation, and
causing other dependencies to end up in a broken state. Solution here is
to pin `torchmetrics` inside the `syntheseus-chemformer` package.