Correctly specify dependency on typing_extensions (#107)
Syntheseus previously depended on typing_extensions without specifying
this in the `pyproject.toml` file, causing a fresh install to fail the
tests. This PR fixes this by:

1. Wrapping the second import of typing_extensions in an if/else block
to only run if python < 3.8 is used (otherwise an import from `typing`
can be used directly). This was already present in the first import of
typing_extensions.
2. Modify pyproject.toml to specify that `typing_extensions` is
required, but _only if python < 3.8_.

---------

Co-authored-by: Austin T <austin.james.tripp[at]gmail.com>