Install `torch_scatter` from conda instead of pip (#110)
As highlighted by DocMinus in #106, our installation instructions can
fail due to issues with `torch_scatter`. This is because our current
setup is essentially equivalent to `pip install torch-scatter`, which
compiles the library from source, and so depends on having the right
versions of `nvcc` and `gcc`. Instead, this PR shifts the
`torch_scatter` dependency to use `pyg`'s conda channel as opposed to
installation through `pip`.